requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
